[IMP] test_mass_mailing: improve test data management and test users, merge mailing tests
In this commit we try to simplify tests organization by

  * merging mailing_internals and mass_mailing tests, as they globally handle
    mailing-related tests (blacklist, opt-out, gateway update through replies
    and clicks);
  * move wizard (mailing.test) tests in their own file as they target a
    specific model;
  * add some loggers to reduce noise;
  * add some users wrappers to test using real users;

Backport of odoo/odoo@230b8258a551010080dd582923afe5a2f91d644b done in master (14.4)

LINKS

Task ID-2117639
COM PR odoo/odoo#40931

X-original-commit: 4cf75bb1767fe7524c8e4d895d84ad2b56ef404d
