+++ addons/test_mass_mailing/tests/__init__.py
@@ -5,6 +5,6 @@ from . import test_blacklist
 from . import test_blacklist_behavior
 from . import test_blacklist_mixin
 from . import test_link_tracker
-from . import test_mailing_internals
-from . import test_mass_mailing
+from . import test_mailing_test
+from . import test_mailing
 from . import test_performance
